requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.1
httplib2==0.9.2
PyMySQL==0.9.3
razorpay==1.1.1
urllib3==1.24.1
requests==2.21.0

boto3==1.9.82
botocore==1.12.82
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
idna==2.8
jmespath==0.9.3
python-dateutil==2.7.5
pytz==2018.9
razorpay==1.1.1
requests==2.21.0
rijndael==0.1.1
s3transfer==0.1.13
six==1.12.0
typesense==0.3.0